cua-driver: address CodeRabbit nits on MCP TCC auto-delegation
- Extract the "is this binary running from inside an installed
  CuaDriver.app bundle?" heuristic into a single shared
  `isExecutableInsideCuaDriverApp()` helper in
  `CuaDriverCLI/BundleHelpers.swift`. `ServeCommand` and `MCPCommand`
  both call into it now, instead of carrying byte-identical local
  copies. Subcommands can still wrap it with extra env/flag/ppid
  gating where their relaunch heuristics diverge.

- Make `CuaDriverMCPServer.fetchProxyToolList` (and therefore
  `makeProxy`) throwing. Previously a missing/unhealthy daemon
  silently returned an empty tool list, so the MCP client saw a
  "successful" handshake advertising zero tools and then errored on
  every subsequent `CallTool`. Now it throws a descriptive
  `MCPError.internalError` pointing at the socket path and the
  `open -n -g -a CuaDriver --args serve` recovery, which surfaces
  during proxy init and gets logged to stderr by `AppKitBootstrap`'s
  catch path before the process exits.
